Rename helpers.ts to tmp-dir.ts
