+++ extensions/ql-vscode/src/query-server/legacy/run-queries.ts
@@ -9,7 +9,7 @@ import {
   DatabaseItem,
   DatabaseResolver,
 } from "../../databases/local-databases";
-import { tmpDir } from "../../helpers";
+import { tmpDir } from "../../tmp-dir";
 import { ProgressCallback } from "../../common/vscode/progress";
 import { QueryMetadata } from "../../pure/interface-types";
 import { extLogger, Logger } from "../../common";
