+++ extensions/ql-vscode/src/tmp-dir.ts
