+++ extensions/ql-vscode/test/vscode-tests/no-workspace/query-results.test.ts
@@ -20,7 +20,7 @@ import {
 } from "../../../src/pure/interface-types";
 import { CodeQLCliServer, SourceInfo } from "../../../src/codeql-cli/cli";
 import { CancellationTokenSource, Uri } from "vscode";
-import { tmpDir } from "../../../src/helpers";
+import { tmpDir } from "../../../src/tmp-dir";
 import {
   formatLegacyMessage,
   QueryInProgress,
